+++ lib/src/spot/text/any_text.dart
@@ -237,9 +237,7 @@ class AnyTextWidgetSelector extends WidgetSelector<AnyText> {
   /// - `children`: Child selectors to include in the match.
   /// - `parents`: Parent selectors to include in the match.
   AnyTextWidgetSelector({
-    required super.props,
-    super.children,
-    super.parents,
+    required super.stages,
   }) : super(mapElementToWidget: _mapElementToAnyText);
 
   static AnyText _mapElementToAnyText(Element element) {
@@ -256,13 +254,6 @@ class AnyTextWidgetSelector extends WidgetSelector<AnyText> {
       'Widget ${element.widget.toStringShort()} is not supported by AnyText',
     );
   }
-
-  @override
-  List<ElementFilter> createElementFilters() {
-    return super.createElementFilters()
-      // Matches multiple widget types, can't filter by synthetic type AnyText
-      ..removeWhere((it) => it is WidgetTypeFilter);
-  }
 }
 
 /// Matches text widgets ([EditableText] and [RichText]) on screen.
@@ -270,13 +261,13 @@ class AnyTextWidgetSelector extends WidgetSelector<AnyText> {
 /// This predicate is used to verify the presence and properties of text within
 /// widgets. It extracts text data from the relevant widget and uses the
 /// provided [match] function to assert the text content.
-class MatchTextPredicate implements PredicateWithDescription {
-  /// Constructs a [MatchTextPredicate].
+class MatchTextFilter implements ElementFilter {
+  /// Constructs a [MatchTextFilter].
   ///
   /// The [match] function is used to assert the text content found in the
   /// relevant widget. The [description] provides a human-readable explanation
   /// of what this predicate checks.
-  MatchTextPredicate({
+  MatchTextFilter({
     required this.match,
     required this.description,
   });
@@ -290,7 +281,9 @@ class MatchTextPredicate implements PredicateWithDescription {
   final String description;
 
   @override
-  bool Function(Element e) get predicate => _match;
+  Iterable<WidgetTreeNode> filter(Iterable<WidgetTreeNode> candidates) {
+    return candidates.where((it) => _match(it.element));
+  }
 
   bool _match(Element element) {
     try {