+++ test/spot/existence_comparison_test.dart
@@ -129,7 +129,7 @@ void main() {
       final material = spot<Material>(parents: [spot<SizedBox>()]);
 
       final throwsFailureWithMessage = throwsSpotErrorContaining(
-        ["Could not find SizedBox > Material in widget tree"],
+        ["Could not find SizedBox ᗕ Material in widget tree"],
       );
 
       expect(material.snapshot().discovered, isEmpty);
@@ -164,10 +164,10 @@ void main() {
       expect(selector.snapshot().discovered, isEmpty);
 
       final throwsFailureWithMessage = throwsSpotErrorContaining([
-        "Could not find Material > Center with children: [SizedBox] in widget tree, expected",
+        "Could not find Material ᗕ Center with child SizedBox in widget tree, expected",
         RegExp(r"(?:exactly|at most|at least) \d+"),
         RegExp(
-          r"A less specific search \(Center with children: \[SizedBox\]\) discovered \d+ matches",
+          r"A less specific search \(Center with child SizedBox\) discovered \d+ matches",
         ),
         "Possible match #1:\nCenter(alignment: Alignment.center",
       ]);
@@ -259,9 +259,9 @@ void main() {
         () => spot<Text>(parents: [spot<Row>()]).existsAtLeastNTimes(2),
         throwsSpotErrorContaining(
           [
-            "Found 1 elements matching Row > Text in widget tree, expected at least 2",
+            "Found 1 elements matching Row ᗕ Text in widget tree, expected at least 2",
             "A less specific search (Text) discovered 3 matches!",
-            "Maybe you have to adjust your WidgetSelector (Text with parents: [Row]) to cover those missing elements.",
+            "Maybe you have to adjust your WidgetSelector (Text with parent Row) to cover those missing elements.",
             'Possible match #1:\nText("a"',
             'Possible match #2:\nText("b"',
             'Possible match #3:\nText("c"',