+++ test/spot/exists_once_test.dart
@@ -108,7 +108,7 @@ void main() {
       throwsSpotErrorContaining(
         [
           'Found 2 elements',
-          "Wrap > Text",
+          "Wrap ᗕ Text",
           'expected exactly 1',
           '\nWrap(', // at the beginning of the line, common ancestor
           'Text("World"',
@@ -211,7 +211,7 @@ void main() {
     final textSnapshot = spot<Text>().snapshot();
     expect(textSnapshot.discoveredElements.length, 2);
 
-    final wrap = spot<Wrap>()..snapshot().existsOnce();
+    final wrap = spot<Wrap>();
     // only find the single sizedBox below Wrap
     wrap.spot<SizedBox>().existsOnce();
 
@@ -220,11 +220,7 @@ void main() {
 
     snapshot(spot<Text>()).existsAtLeastOnce();
 
-    multipleSpotter.copyWith(
-      parents: [
-        // only finds the single SizedBox in Wrap, not the SizedBox below Center
-        wrap.spot<SizedBox>(),
-      ],
-    ).existsOnce();
+    // only finds the single SizedBox in Wrap, not the SizedBox below Center
+    multipleSpotter.withParent(wrap.spot<SizedBox>()).existsOnce();
   });
 }